moved the envvars locally

+++ files/run.sh
@@ -18,7 +18,11 @@ cp elemental_static-ifs  /root/bin/elemental_static-ifs
 
 fi
 
-cp set-static /root/bin/set-static
+if [ ! -f /root/bin/set-static || -f /root/bin/static.env ]
+then
+	cp set-static /root/bin/set-static
+	cp static.env /root/bin/static.env
+fi
 
 if [ -f /etc/rancher/static/STATIC-FAILED ]; then
 	systemctl disable --now --no-block elemental-static-network
